requirements.txt
flask==3.1.2
flask-cors==6.0.1
requests==2.31.0
python-dotenv==1.0.0
